Refactoring block_for_gc (#100)
Currently, we keep calling into C multiple times to set up some
variables before the thread stops for GC and after GC is done. This PR
refactors the code in `block_for_gc` to delegate that to C, reusing most
of the code in `jl_gc_collect`. That should help preventing errors in
the future due to the mismatch between the implementations as it keeps
things in a single place.

Should be merged with https://github.com/mmtk/julia/pull/29.