Remove code that iterates over bindings table (#91)
This PR tries to remove code that iterates over bindings table within a
module. This may cause issues because we load information from this
table, and if/when we support moving, the fields loaded might be
obsolete. Given that we process the edge `&m->bindings`, it is not clear
whether we need to iterate over the table when processing the module.
Perhaps, before merging this PR we should consult with people familiar
with Julia's stock GC to confirm that this change won't cause any
unknown problems.
Note that this PR should close
https://github.com/mmtk/mmtk-julia/issues/49.