Stop using Julia's size classes when using MMTk (#108)
When allocating an object, we are still allocating the number of bytes
according to Julia's size classes, which may be an overestimation of the
number of bytes an object actually requires. This PR removes this since
this is specific to Julia's stock GC. This affects not only allocation
itself but needs to be reflected in the `get_current_size` function.

Needs to be merged with https://github.com/mmtk/julia/pull/31.

(cherry picked from commit efb9e105f830c1644b735b6fd24774914408075e)

# Conflicts:
#	mmtk/Cargo.toml