Process GC preserve (#156)
Processing GC preserve regions. 
Note that objects in these regions are currently put into a separate
stack (`tpin_gcstack`), but later on, when we lower the code for pushing
and popping objects into LLVM, we might be able to reuse the `gcstack`
and create a separate frame that we can push and pop. We cannot
currently do that because we are currently pushing and popping outside
of the current function, using the hook functions.

---------

Co-authored-by: Luis Eduardo de Souza Amorim <eduardo@bear.moma>