Conservative stack scanning (#157)
This uses conservative stack scanning to find object references from the
execution stack, and pins those values. The changes in the PR include:
* Conservatively find object references from stacks and registers, and
pin those references. The references are unpinned after the GC if they
are still alive.
* Calling `jl_save_context_for_conservative_scanning` for the first
thread that is blocked for the GC.
* Allow getting object size for large objects.
* Refactoring and cleaning up features. We should choose either the
`moving` or the `non_moving` feature.

---------

Co-authored-by: Eduardo Souza <ledusou@gmail.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>