README.md: expand build instructions and put "quick start" first (#195)
This expands the build instructions to mention a few things previously
missing, including the need to set JULIA_PATH and to check out the 'dev'
branch. Without the latter, you will get a working Julia but it is not
MMTk-enabled.

I also tried to eliminate a bit of duplication, have the "quick build"
instructions be a sequence of verbatim shell commands that Should Work
(tested working for me) and put them first.