Add an feature to record moved objects (#258)
Add a debug feature to record information about moved objects. When we
move an object, we record the object's old reference (which becomes
dangling soon), and its type in a map. Then we see the dangling pointer,
we can query the map to know its original type. We could expand this to
record more information other than the type. This could be helpful for
debugging.